Make a response that looks like the request format

+++ socket_class.h
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <optional>
 #include <expected>
 #include <exception>
@@ -37,7 +38,8 @@ public:
     // get the next msg from the kernel
     std::optional<std::string> get_msg() noexcept; 
 
-    int send_response(int statuscode, char *buff, int len);
+    int send_response(uint8_t statuscode, char *buff, int len);
+
 
     void read_msg();
